Rework opening times
Reworked design of opening_times on client side so that they can be populated when redirecting back if a validation error occurs, or editing an existing casino.
This new design meant the middleware for transforming casino opening times is no longer needed as the form fields are already in the desired format.
PROTIP: For arrays of inputs use group[KEY][INPUT] instead of group[]INPUT