Present opening times in user friendly format
The data for opening times are stored as a true representation. For
example, a casino opening from Monday 16:00pm to Tuesday 04:00 am has 2
opening times stored. Monday 16:00 to 23:59 and Tuesday 00:00 to 04:00.

Methods have been implemented to convert a casinos opening times from
the true representation to a more user friendly format.